testapp debug.c, cleanup loader debug.c
